requirements.txt
alembic==1.15.1
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.39
typeguard==4.4.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
